refactor

+++ crates/crabe_decision/src/strategy/testing/circlerotation.rs
@@ -48,7 +48,7 @@ impl Strategy for CircleRotation {
         let elapsed = (self.start_time.elapsed().as_millis() as f64)/1000.;
         let mut itt: f64 = 0.;
         for id in self.ids.iter(){
-            let relative_elapsed = elapsed  + (((PI*2./self.ids.len() as f64) * itt) as f64);
+            let relative_elapsed = elapsed  + (PI*2./self.ids.len() as f64) * itt;
             dbg!(relative_elapsed, id);
             let sin = f64::sin(relative_elapsed);
             let cos = f64::cos(relative_elapsed);
